Adapted to interface change in OTDet/RtRelation
git-svn-id: http://svn.cern.ch/guest/lhcb/Alignment/trunk@185499 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 34c91f3755c2ec26285b9e19cb29b2322d3f3c85

+++ Alignment/Calibration/OTCalibration/src/OTModuleClbrAlg.h
@@ -132,8 +132,8 @@ private:
    */
   bool fitTR;
 
-  std::vector<double> tcoeff;
-  std::vector<double> terrcoeff;
+  std::vector<float> tcoeff;
+  std::vector<float> terrcoeff;
 
   double t0s[3][4][4][9];
